+++ lib/puppet_x/puppetlabs/pelookup.rb
@@ -1,22 +1,18 @@
-#!/opt/puppetlabs/puppet/bin/ruby
-
-# Output a class parameter defined in Hiera and/or the Classifier.
+# Output a key defined in Hiera and/or the Classifier.
 
 module PuppetX
   module Puppetlabs
-    # Output a class parameter defined in Hiera and/or the Classifier.
-    class Lookup
+    # Output a key defined in Hiera and/or the Classifier.
+    class PELookup
       attr_reader :environment
+      attr_reader :node
 
       # Initialize command options, class variables, and objects.
       def initialize(options)
-        unless options[:param]
-          output_error_and_exit('The --param option is required')
-        end
-
         unless options[:node]
           output_error_and_exit('The --node option is required')
         end
+        @node = options[:node]
 
         unless options[:pe_environment]
           output_error_and_exit('The --pe_environment option is required')
@@ -24,40 +20,42 @@ module PuppetX
         @environment = options[:pe_environment]
       end
 
-      # Output the specified class parameter defined in Hiera and/or the Classifier.
+      # Output the specified key defined in Hiera and/or the Classifier.
 
-      def output_current_setting(certname, setting_name)
-        current_settings = hiera_classifier_settings(certname, [setting_name])
+      def lookup(setting_name)
+        output_error_and_exit('A key to lookup is a required argument') unless setting_name
+
+        current_settings = hiera_classifier_settings(@node, [setting_name])
         output_error_and_exit('Unable to query Hiera or the Classifier') if current_settings.nil?
 
         current_settings['hiera'] = current_settings['hiera'].select { |k, _v| k == setting_name }
         current_settings['classifier'] = current_settings['classifier'].select { |k, _v| k == setting_name }
-        output 'Node: %{certname}' % { certname: certname }
-        output 'Parameter: %{setting_name}' % { setting_name: setting_name }
+        output 'Node: %{node}' % { node: @node }
+        output 'Key:  %{setting_name}' % { setting_name: setting_name }
         output_line
 
         found_in_hiera = current_settings['hiera'].key?(setting_name)
         if found_in_hiera
-          output 'Parameter found in Hiera:'
+          output 'Key found in Hiera:'
           output_line
           output_data(current_settings['hiera'].to_yaml)
         else
-          output 'Parameter not found in Hiera'
+          output 'Key not found in Hiera'
         end
         output_line
 
         found_in_classifier = current_settings['classifier'].key?(setting_name)
         if found_in_classifier
-          output 'Parameter found in the Classifier:'
+          output 'Key found in the Classifier:'
           output_line
           output_data JSON.pretty_generate(current_settings['classifier'])
         else
-          output 'Parameter not found in the Classifier'
+          output 'Key not found in the Classifier'
         end
         output_line
 
         return unless found_in_hiera && found_in_classifier
-        output 'Classifier settings take precedence over Hiera settings.'
+        output 'The Classifier take precedence over Hiera.'
         output_line
       end
 
@@ -67,9 +65,9 @@ module PuppetX
 
       def hiera_classifier_settings(certname, setting_names)
         overrides_hiera, overrides_classifier = hiera_classifier_overrides(certname, setting_names)
-        Puppet.debug("Settings: #{setting_names}")
-        Puppet.debug("Settings from Hiera for: #{certname}: #{overrides_hiera}")
-        Puppet.debug("Settings from Classifier for: #{certname}: #{overrides_classifier}")
+        Puppet.debug("Keys: #{setting_names}")
+        Puppet.debug("Keys from Hiera for: #{certname}: #{overrides_hiera}")
+        Puppet.debug("Keys from Classifier for: #{certname}: #{overrides_classifier}")
         return { 'hiera' => overrides_hiera, 'classifier' => overrides_classifier }
       rescue Puppet::Error
         return nil
@@ -145,5 +143,5 @@ end
 # or as 'puppet pe lookup'
 
 if File.expand_path(__FILE__) == File.expand_path($PROGRAM_NAME)
-  require_relative 'lookup/cli'
+  require_relative 'pelookup/cli'
 end