+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.5.0
+
+- convert '--param' from an option to an argument
+- rename class Lookup to PELookup
+- normalize all 'setting' and 'class parameter' references to 'key'
+
 ## 1.0.2
 
 - Remove and move puppet/util/pe_conf